+ Float input GUI view  + Vertices are now double precision (needed to) 	* When using low DT, inner planets would move, while outer planets wasn't  + Body rotation 	* Hardcoded rotation value at the moment, and performned on the CPU (for testing)

+++ include/BodyCameraControl.hpp
@@ -17,10 +17,10 @@ class BodyCameraControl: public AbstractCamera{
 		Body *body;
 		
 		//Camera data
-		float distance;
-		float horizontalAngle;
-		float verticalAngle;
-		float mouseSpeed;
+		double distance;
+		double horizontalAngle;
+		double verticalAngle;
+		double mouseSpeed;
 		double previousTime;
 	public:
 		/**
